requirements.txt
absl-py
altgraph
anyio
argon2-cffi
argon2-cffi-bindings
asgiref
asttokens
astunparse
attrs
backcall 
cachetools
certifi
cffi
charset-normalizer
click
colorama
commonmark
crispy-bootstrap5
cryptography
cycler
dacite
debugpy
decorator
defusedxml
Deprecated
diff-match-patch
distlib
dj-database-url
Django
django-bootstrap-modal-forms
django-bootstrap3
django-crispy-forms
django-embed-video
django-environ
django-filter
django-forms-bootstrap
django-import-export
django-widget-tweaks
djangorestframework
docutils
entrypoints
et-xmlfile
exceptiongroup
executing 
flatbuffers
fonttools
gast
gevent
google-auth
google-auth-oauthlib
google-pasta
greenlet
grequests
grpcio
gunicorn
h11
h5py
html5lib
htmlmin
idna
ImageHash
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jaraco.classes
jedi
Jinja2
joblib
jsonschema
keras
keyring
libclang
lxml
Markdown
MarkupPy
MarkupSafe
mediapipe
mistune
more-itertools
MouseInfo
mssql-django
multimethod
nbclient
nbconvert
nbformat
nest-asyncio
networkx
oauthlib
odfpy
opt-einsum
outcome
packaging
pandocfilters
parso
patsy
pefile
phik
pickleshare
pillow
pipenv
pkginfo
platformdirs
prometheus-client
prompt-toolkit
protobuf
psutil
psycopg2
psycopg2-binary
pure-eval
pyasn1
pyasn1-modules
pycparser
pycryptodome
pydantic 
PyGetWindow
PyGithub
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyJWT
pyloco
PyMsgBox
PyMuPDF
PyMuPDFb
PyNaCl
pyodbc
pyOpenSSL
pyparsing
PyPDF2
pyperclip
pypiwin32
PyRect
pyrsistent
PyScreeze
PySimpleGUI
PySocks
python-barcode
python-dateutil
python-decouple
python-docx
python-dotenv
python-multipart
pytweening
pytz
PyWavelets
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
readme-renderer
requests
requests-oauthlib
requests-toolbelt
rfc3986
rich
rsa
Send2Trash
SimpleWebSocketServer
six
sniffio
sortedcontainers
soupsieve
sqlparse
stack-data
starlette
statsmodels
tablib
tangled-up-in-unicode
termcolor
terminado
testpath
threadpoolctl
tornado
tqdm
traitlets
trio
trio-websocket
twine
typeguard
typing
typing_extensions
tzdata
urllib3
ushlex
uvicorn
virtualenv
visions
wcwidth
webdriver-manager
webencodings
websocket-client
Werkzeug
whitenoise
widgetsnbextension
wordcloud
wrapt
wsproto
xlrd
xlwt
ydata-profiling
zipp
zope.event
zope.interface